+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add a refresh button to the SQL console schema tree.
+
 ## 2024-11-29 - 0.19.2
 
 - Dependabot: bump antd from 4.24.16 to 5.22.2.